# docs/swagger.json
{
  "components": {
...
    "schemas": {
...
      "Article": {
        "properties": {
          "content": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string"
          },
          "urlToImage": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "source": {
            "properties": {
              "name": {
                "type": "string"
              },
              "id": {
                "type": "string"
              }
            },
            "required": ["name", "id"],
            "type": "object"
          }
        },
        "required": [
          "content",
          "publishedAt",
          "urlToImage",
          "url",
          "description",
          "title",
          "author",
          "source"
        ],
        "type": "object"
      },
      "SearchIn": {
        "type": "string",
        "enum": ["title", "description", "content", ""]
      },
      "Language": {
        "type": "string",
        "enum": [
          "ar",
          "de",
          "en",
          "es",
          "fr",
          "he",
          "it",
          "nl",
          "no",
          "pt",
          "ru",
          "sv",
          "ud",
          "zh",
          ""
        ]
      },
      "SortBy": {
        "type": "string",
        "enum": ["relevancy", "popularity", "publishedAt"]
      },
      "CountryCode": {
        "type": "string",
        "enum": [
          "ae",
          "ar",
          "at",
          "au",
          "be",
          "bg",
          "br",
          "ca",
          "ch",
          "cn",
          "co",
          "cu",
          "cz",
          "de",
          "eg",
          "fr",
          "gb",
          "gr",
          "hk",
          "hu",
          "id",
          "ie",
          "il",
          "in",
          "it",
          "jp",
          "kr",
          "lt",
          "lv",
          "ma",
          "mx",
          "my",
          "ng",
          "nl",
          "no",
          "nz",
          "ph",
          "pl",
          "pt",
          "ro",
          "rs",
          "ru",
          "sa",
          "se",
          "sg",
          "si",
          "sk",
          "th",
          "tr",
          "tw",
          "ua",
          "us",
          "ve",
          "za",
          ""
        ]
      },
      "Category": {
        "type": "string",
        "enum": [
          "business",
          "entertainment",
          "general",
          "health",
          "science",
          "sports",
          "technology",
          ""
        ]
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "apis",
    "version": "0.0.1",
    "description": "APIs",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "name": "hieudoanm ",
      "email": "hieumdoan@gmail.com"
    }
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/news": {
      "get": {
        "operationId": "GetNews",
        "responses": {
          "200": {
            "description": "Get News",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["News"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "searchIn",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SearchIn"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Language"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortBy"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 100,
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
    "/api/news/headlines": {
      "get": {
        "operationId": "GetHeadlines",
        "responses": {
          "200": {
            "description": "Get Headlines",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["News"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CountryCode"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Category"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 100,
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}